feat: sync computed camera values

+++ src/math/Mat3.ts
@@ -363,7 +363,7 @@ export class Mat3 {
 }
 
 export const m3Type = Type.vector(
-  Type.float64,
+  Type.float32,
   ['m00', 'm01', 'm02', 'm10', 'm11', 'm12', 'm20', 'm21', 'm22'],
   // @ts-ignore
   Mat3,
